Fixed opcode definitions, added asuka1
Big update, I've changed a number of opcode definitions (check the sanny
data folder) and updated all the files to use them. This also fixed some
bugs caused by wrong defintions before.
NOTE: because of the changed definitions, these files cannot be mixed
with the older ones, when using old files, the opcodes with a changed
parameter order NEED to be updated to the new order to compile properly
with the new SCM.ini (in the same way the new files wont compile
correctly with the old SCM.ini). All opcodes with a changed parameter
order are noted in the sanny data folder.

Added asuka1, AI handling in that mission is a pain.

Also fixed some smaller bugs here and there which I can't remember
exactly.