Created automatic compilation .cmd
As a tribute to laziness, I have made _compile.cmd which when run will
compile main.sc and main_debug.sc automatically to the Compiled SCM
folder.
debug.sc now sets $MASTERDEBUG = 0 instead of 1, main_debug overrides
this and sets it to 1.
The debug main is renamed but works exactly the same.
_compile.cmd tries to get the Sanny Builder install directory from the
registry, if this fails you can easily edit the .cmd file and set it
manually.